Remove code duplications in test
